restructured deployment files

+++ src/main/kotlin/studio/pinkcloud/voyager/deployment/DeploymentSystemImpl.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.encodeToString
 import studio.pinkcloud.voyager.VOYAGER_JSON
 import studio.pinkcloud.voyager.deployment.caddy.ICaddyManager
 import studio.pinkcloud.voyager.deployment.cloudflare.ICloudflareManager
-import studio.pinkcloud.voyager.deployment.data.Deployment
+import studio.pinkcloud.voyager.deployment.data.*
 import studio.pinkcloud.voyager.deployment.discord.IDiscordManager
 import studio.pinkcloud.voyager.deployment.docker.IDockerManager
 import studio.pinkcloud.voyager.utils.Env
@@ -12,7 +12,6 @@ import studio.pinkcloud.voyager.utils.PortFinder
 import java.io.File
 
 class DeploymentSystemImpl : IDeploymentSystem {
-
     private val deployments: MutableList<Deployment> = mutableListOf()
     private val deploymentsFile = File("/opt/pinkcloud/voyager/deployments.json")
 
@@ -20,8 +19,8 @@ class DeploymentSystemImpl : IDeploymentSystem {
         if (deploymentsFile.exists()) {
             deployments.addAll(
                 VOYAGER_JSON.decodeFromString(
-                    deploymentsFile.readText()
-                )
+                    deploymentsFile.readText(),
+                ),
             )
         } else {
             deploymentsFile.createNewFile()
@@ -33,16 +32,19 @@ class DeploymentSystemImpl : IDeploymentSystem {
         Runtime.getRuntime().addShutdownHook(
             Thread {
                 deploymentsFile.writeText(
-                    VOYAGER_JSON.encodeToString(deployments)
+                    VOYAGER_JSON.encodeToString(deployments),
                 )
                 ICaddyManager.INSTANCE.updateCaddyFile(getCaddyFileContent(), withOurApi = false)
-            }
+            },
         )
     }
 
-    override suspend fun deploy(deploymentKey: String, dockerFile: File): String {
+    override suspend fun deploy(
+        deploymentKey: String,
+        dockerFile: File,
+    ): String {
         // call deployment functions in IDockerManager [Done]
-        // notify discord bot. 
+        // notify discord bot.
         // add to caddy. [Done]
         // add to deployment list  [Done]
         // add to cloudflare dns. [Done]
@@ -54,15 +56,24 @@ class DeploymentSystemImpl : IDeploymentSystem {
         IDockerManager.INSTANCE.buildDockerImage(deploymentKey, dockerFile)
 
         val port = PortFinder.findFreePort() // the port that the reverse proxy needs to use.
-        val containerId = IDockerManager.INSTANCE.createAndStartContainer(deploymentKey, port, findInternalDockerPort(dockerFile), deploymentKey)
+
+        // TODO: check for failed deployment
+        val containerId =
+            IDockerManager.INSTANCE.createAndStartContainer(
+                deploymentKey,
+                port,
+                findInternalDockerPort(dockerFile),
+                deploymentKey,
+            )
 
         deployments.add(
             Deployment(
                 deploymentKey,
                 port,
                 containerId,
-                cloudflareId
-            )
+                cloudflareId,
+                DeploymentState.DEPLOYED
+            ),
         )
 
         // add to caddy.
@@ -70,34 +81,46 @@ class DeploymentSystemImpl : IDeploymentSystem {
 
         // notify discord bot.
         IDiscordManager.INSTANCE.sendDeploymentMessage(deploymentKey, port, containerId)
-        
+
         return containerId
     }
 
-    override suspend fun stop(deployment: Deployment) {
-        // stop docker container.
-        IDockerManager.INSTANCE.stopContainerAndDelete(deployment.dockerContainer)
+    override suspend fun stopAndDelete(deployment: Deployment) {
+        stop(deployment)
         delete(deployment)
+    }
+
+    override suspend fun stop(deployment: Deployment) {
+        // stop docker container
+        if (deployment.state != DeploymentState.DEPLOYED) return
+        deployment.state = DeploymentState.STOPPING
+        IDockerManager.INSTANCE.stopContainer(deployment.dockerContainer);
+        deployment.state = DeploymentState.STOPPED
     }
 
     override suspend fun delete(deployment: Deployment) {
         // stop and remove docker container.
+        if (deployment.state != DeploymentState.STOPPED) return
+        deployment.state = DeploymentState.DELETING
+        IDockerManager.INSTANCE.deleteContainer(deployment.dockerContainer)
 
         // remove any existing files.
-        File("/opt/pinkcloud/voyager/deployments/${deployment.deploymentKey}-preview").also { 
+        File("/opt/pinkcloud/voyager/deployments/${deployment.deploymentKey}-preview").also {
             if (it.exists()) {
                 it.deleteRecursively()
             }
         }
-        
+
         // remove from deployment list [done]
         deployments.remove(deployment)
 
         // remove from caddy after it is removed from internals deployments list. [done]
         ICaddyManager.INSTANCE.updateCaddyFile(getCaddyFileContent())
-        
+
         // remove from cloudflare dns.[done]
         ICloudflareManager.INSTANCE.removeDnsRecord(deployment.deploymentKey)
+
+        deployment.state = DeploymentState.DELETED
     }
 
     override fun getLogs(deployment: Deployment): String {
@@ -107,13 +130,14 @@ class DeploymentSystemImpl : IDeploymentSystem {
     override fun getCaddyFileContent(): String {
         var content = ""
 
-        deployments.forEach {  deployment ->
-            content += """
-               
-               ${deployment.deploymentKey}-preview.pinkcloud.studio {
-                   reverse_proxy localhost:${deployment.port}
-               }
-            """.trimIndent()
+        deployments.forEach { deployment ->
+            content +=
+                """
+                
+                ${deployment.deploymentKey}-preview.pinkcloud.studio {
+                    reverse_proxy localhost:${deployment.port}
+                }
+                """.trimIndent()
         }
 
         return content
@@ -129,5 +153,15 @@ class DeploymentSystemImpl : IDeploymentSystem {
 
     private fun findInternalDockerPort(dockerFile: File): Int {
         return dockerFile.readText().substringAfter("EXPOSE ").substringBefore("\n").toInt()
+    }
+
+    override fun isRunning(deployment: Deployment): Boolean {
+        if (deployment.state != DeploymentState.DEPLOYED) return false
+        return IDockerManager.INSTANCE.isContainerRunning(deployment.dockerContainer)
+    }
+
+    override suspend fun restart(deployment: Deployment) {
+        if (deployment.state == DeploymentState.DEPLOYED) stopAndDelete(deployment)
+        if (deployment.state != DeploymentState.STOPPED) return
     }
 }
